Modified

+++ Repositories/UserRepository.cs
@@ -2,12 +2,16 @@
 using HotPotProject.Interfaces;
 using HotPotProject.Models;
 using Microsoft.EntityFrameworkCore;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
 
 namespace HotPotProject.Repositories
 {
-    public class UserRepository: IRepository<int, string, User>
+    public class UserRepository : IRepository<int, string, User>
     {
-        ApplicationTrackerContext _context;
+        private readonly ApplicationTrackerContext _context;
 
         public UserRepository(ApplicationTrackerContext context)
         {
@@ -17,42 +21,45 @@ namespace HotPotProject.Repositories
         public async Task<User> Add(User item)
         {
             _context.Add(item);
-            _context.SaveChanges();
+            await _context.SaveChangesAsync();
             return item;
         }
 
         public async Task<User> Delete(int key)
         {
-            var user = await GetAsync(key);
-            _context.Users.Remove(user);
-            _context.SaveChanges();
+            var user = await _context.Users.FindAsync(key);
+            if (user != null)
+            {
+                _context.Users.Remove(user);
+                await _context.SaveChangesAsync();
+            }
             return user;
         }
 
-        public Task<User> GetAsync(int key)
+        public async Task<User> GetAsync(int key)
         {
-            throw new NotImplementedException();
+            return await _context.Users.FindAsync(key);
         }
 
         public async Task<List<User>> GetAsync()
         {
-            var users = _context.Users.ToList();
-            return users;
+            return await _context.Users.ToListAsync();
         }
 
         public async Task<User> GetAsync(string key)
         {
-            var users = await GetAsync();
-            var user = users.FirstOrDefault(u => u.UserName == key);
-            return user;
+            return await _context.Users.FirstOrDefaultAsync(u => u.UserName == key);
         }
 
         public async Task<User> Update(User item)
         {
-            var user = await GetAsync(item.UserName);
-            _context.Entry<User>(item).State = EntityState.Modified;
-            _context.SaveChanges();
-            return item;
+            var existingUser = await _context.Users.FindAsync(item.UserName);
+            if (existingUser != null)
+            {
+                _context.Entry(existingUser).CurrentValues.SetValues(item);
+                await _context.SaveChangesAsync();
+            }
+            return existingUser;
         }
     }
 }
